Kill some warnings.
git-svn-id: http://www.telegraphics.com.au/svn/psdparse/trunk@277 d127708d-35f1-0310-882e-b6c776866be2

+++ util.c
@@ -90,7 +90,7 @@ void fputcxml(wchar_t c, FILE *f){
 		if(c < 0x80 && isprint(c)) // ASCII printable
 			fputc(c, f);
 		else
-			fprintf(f, "&#x%04x;", c);
+			fprintf(f, "&#x%04x;", (unsigned)c);
 	}
 }
